+++ ProgramFunctions/Update.fnc.php
@@ -376,7 +376,7 @@ function _update69beta()
  * Update to version 8.1
  *
  * 1. accounting_salaries table: Add FILE_ATTACHED column.
- * 2. BILLING_FEES table: Add FILE_ATTACHED column.
+ * 2. billing_fees table: Add FILE_ATTACHED column.
  *
  * Local function
  *
@@ -404,7 +404,7 @@ function _update81()
 	}
 
 	/**
-	 * 2. BILLING_FEES table: Add FILE_ATTACHED column.
+	 * 2. billing_fees table: Add FILE_ATTACHED column.
 	 */
 	$file_attached_column_exists = DBGetOne( "SELECT 1 FROM pg_attribute
 		WHERE attrelid=(SELECT oid FROM pg_class WHERE relname='billing_fees')