+++ modules/Student_Billing/MassAssignFees.php
@@ -16,7 +16,7 @@ if ( $_REQUEST['modfunc'] === 'save' )
 
 			foreach ( (array) $_REQUEST['student'] as $student_id )
 			{
-				$sql .= "INSERT INTO BILLING_FEES (STUDENT_ID,TITLE,AMOUNT,SYEAR,SCHOOL_ID,ASSIGNED_DATE,DUE_DATE,COMMENTS)
+				$sql .= "INSERT INTO billing_fees (STUDENT_ID,TITLE,AMOUNT,SYEAR,SCHOOL_ID,ASSIGNED_DATE,DUE_DATE,COMMENTS)
 					VALUES('" . $student_id . "',
 					'" . $_REQUEST['title'] . "','" . preg_replace( '/[^0-9.-]/', '', $_REQUEST['amount'] ) . "',
 					'" . UserSyear() . "','" . UserSchool() . "','" . DBDate() . "','" . $due_date . "',