+++ modules/Student_Billing/includes/Dashboard.inc.php
@@ -48,7 +48,7 @@ if ( ! function_exists( 'DashboardStudentBillingAdmin' ) )
 		// Limit Results to Months between User MP Start & End Date.
 		$fees_RET = DBGet( "SELECT SUBSTRING(CAST(ASSIGNED_DATE AS varchar(10)),1,7) AS YEAR_MONTH,
 			SUM(AMOUNT) AS TOTAL_FEES
-			FROM BILLING_FEES
+			FROM billing_fees
 			WHERE SYEAR='" . UserSyear() . "'
 			AND SCHOOL_ID='" . UserSchool() . "'
 			GROUP BY YEAR_MONTH